requirements.txt
gradio
ultralytics
torch
opencv-python
pillow
